Scheduling final version
The last scheduling push was kind of useless. It was my old idea that I was working on in class. This is the new version I've been working on for 4-5 hours with a calendar to chose dates from, blacking out of times and dates that aren't available, and also a pop up that asks if the information you chose was correct or not. I got the calendar idea and how to integrate it from different sources, but I still tried to integrate it this project my own way.