Nurse/Doctor view patient Records
Earlier the code simply pulled the record of the user logged in, whether it was the patient, doctor or nurse. Now the patient view record functionality only allows him to view his own, while the nurse and doctor can view any patient as long as they have the proper info. This was the major focus of this update. However, small accessibility issues still exist in the way I set it up, so I'll need to take away the access of the nurse to write prescriptions through the redirect accessibility links. I also need to link up the view schedule to the data base to pull info from there in order to finish all the user stories case requirements.